requirements.txt
requests
dnspython
python-dotenv
playwright
pytest
